refactor:generate the config types from the schema refactor:replace assertions with invariant lib
