+++ src/lib/types.ts
@@ -1,3 +1,12 @@
+import {
+  Color,
+  CompensationConfig,
+  Config,
+  LineConfig,
+  NoteConfig,
+  StateConfig,
+} from "./generated/config";
+
 export interface StateHints {
   parent: string | null;
   stereotype: string | null;
@@ -38,25 +47,8 @@ export interface AslChoiceTransitionNode {
   StringEquals?: string;
 }
 
-export type Color = `#${string}`;
-export interface StateConfig {
-  BackgroundColor: Color;
-  FontColor?: Color | "automatic";
-}
 export interface ContainerStateConfig {
   BackgroundColor: Color;
-}
-export interface LineConfig {
-  bold?: true;
-  color: Color;
-}
-export interface CompensationConfig {
-  pattern: RegExp | string;
-  color: Color;
-}
-export interface NoteConfig {
-  width: number;
-  side: "left" | "right";
 }
 
 export interface UserSpecifiedConfig {
@@ -80,31 +72,6 @@ export interface UserSpecifiedConfig {
       Fail?: StateConfig;
       Parallel?: ContainerStateConfig;
       Map?: ContainerStateConfig;
-    };
-  };
-}
-
-export interface Config {
-  theme: {
-    compensation: CompensationConfig;
-    compositeStates: Record<string, string>;
-    comments: Record<string, NoteConfig>;
-    lines: {
-      fromCatch: LineConfig;
-      toFail: LineConfig;
-    };
-    skinparams: {
-      ArrowColor: Color;
-    };
-    states: {
-      Choice: StateConfig;
-      Fail: StateConfig;
-      Map: StateConfig;
-      Parallel: StateConfig;
-      Pass: StateConfig;
-      Succeed: StateConfig;
-      Task: StateConfig;
-      Wait: StateConfig;
     };
   };
 }
